backend/config.py
import os
from dotenv import load_dotenv

# Load environment variables from a .env file in development
load_dotenv()

class Config:
    user = os.environ.get("MYSQL_USER", "root")
    password = os.environ.get("MYSQL_PASSWORD")
    if not password:
        raise ValueError("MYSQL_PASSWORD environment variable is required")

    host = os.environ.get("MYSQL_HOST", "localhost")
    port = os.environ.get("MYSQL_PORT", "3306")
    database = os.environ.get("MYSQL_DATABASE", "demodb")

    # Use PyMySQL instead of mysql-connector-python
    SQLALCHEMY_DATABASE_URI = f"mysql+pymysql://{user}:{password}@{host}:{port}/{database}"
    SQLALCHEMY_TRACK_MODIFICATIONS = False

class DevelopmentConfig(Config):
    DEBUG = True

class ProductionConfig(Config):
    DEBUG = False