backend/requirements.txt
flask
flask-sqlalchemy
mysql-connector-python
python-dotenv
flask-migrate
flask-cors